Major code refactoring

+++ 10_Programmieraufgabe/Lambda.cc
@@ -34,7 +34,7 @@ int main()
 
     // Exercise 2
     std::transform(my_vector.begin(), my_vector.end(), my_result.begin(),
-        [](int val) -> int { return pow(val, 2); });
+        [](int val) -> int { return static_cast<int>(pow(val, 2.0)); });
 
     for (const auto &val : my_result)
     {
